Add integration tests directory
